Move Namespace component to namespace module
This makes more sense.
